+++ arm_chip_pic.h
@@ -1,11 +1,21 @@
-/*!\file arm_chip_ino.h
+/*!\file arm_chip_pic.h
 ** \author SMFSW
 ** \copyright MIT (c) 2017-2022, SMFSW
-** \brief Common macros for Arduino
+** \brief Common macros for PIC
+** \attention	On PIC families you should configure a timer to count for ms.
+**				A TIM peripheral shall be configured in MCC or Harmony (with a period of 1ms).
+**
+** HAL_GetTick shall be known to sarmfsw.
+** As no specific project header is included yet in this file, external definition is provided.
+**
+** \code
+** #include <stdint.h>
+** uint32_t HAL_GetTick(void);
+** \endcode
 */
 /****************************************************************/
-#ifndef __ARM_CHIP_INO_H
-	#define __ARM_CHIP_INO_H
+#ifndef __ARM_CHIP_PIC_H
+	#define __ARM_CHIP_PIC_H
 
 #ifdef __cplusplus
 	extern "C" {
@@ -16,33 +26,29 @@
 #include "arm_errors.h"			// Common errors
 #include "arm_cmsis.h"			// HAL & Drivers (following defined platform)
 /****************************************************************/
-#if ARDUINO > 22
-	#include "Arduino.h"		// Common Arduino definitions
-#else
-	#include "WProgram.h"		// Common Arduino definitions
-#endif
-#include "pins_arduino.h"		// Arduino pins definitions for current board
+extern unsigned int HAL_GetTick(void);	//!< External definition of HAL_GetTick that shall be defined in project
 /****************************************************************/
 
 
 /*** Define Enable/Disable interrupts macros ***/
 #undef diInterrupts
-#define diInterrupts()		noInterrupts()		//!< Disable interruptions macro
+#define diInterrupts()		_GIE = 0;		//!< Disable interruptions macro
 #undef enInterrupts
-#define enInterrupts()		interrupts()		//!< Enable interruptions macro
+#define enInterrupts()		_GIE = 1;		//!< Enable interruptions macro
 
 /*** MS Time base ***/
 #ifndef HAL_MAX_TICKS
-//!\note Define HAL_MAX_TICKS with custom max value in project if tick max value is not using 32b variable full scale
-#define HAL_MAX_TICKS		((uint32_t) -1)		//!< Max Ticks value
+//!\note Define HAL_MAX_TICKS with custom max value in project if tick max value is not using unsigned int variable full scale
+#define HAL_MAX_TICKS		((unsigned int) -1)		//!< Max Ticks value
 #endif
 #ifndef HAL_MS_TICKS_FACTOR
 //!\note Define HAL_MS_TICKS_FACTOR with custom multiplier in project if tick period is not 1ms
 #define HAL_MS_TICKS_FACTOR	1					//!< Milliseconds multiplier (depending tick counter frequency)
 #endif
 
-#define HALTicks()			millis()			//!< Alias for Arduino get ms ticks function
-
+#ifndef HALTicks
+#define HALTicks()			HAL_GetTick()			//!< Alias for Arduino get ms ticks function
+#endif
 
 /*!\enum eResetSource
 ** \brief Source of last reset
@@ -82,5 +88,5 @@ __INLINE FctERR HALERRtoFCTERR(const SDWORD status)
 	}
 #endif
 
-#endif /* __ARM_CHIP_INO_H */
+#endif /* __ARM_CHIP_PIC_H */
 /****************************************************************/