+++ arm_chip_sam.h
@@ -131,6 +131,11 @@ __INLINE FctERR HALERRtoFCTERR(const SDWORD status)
 }
 
 
+#ifndef UNUSED
+#define UNUSED(X)			(void) (X)			//!< To avoid gcc/g++ warnings
+#endif
+
+
 /****************************************************************/
 #ifdef __cplusplus
 	}